+++ textdb/textdb-sandbox/src/main/java/edu/uci/ics/textdb/sandbox/helloworldexamples/AkshayHelloWorld.java
@@ -1,4 +1,4 @@
-package edu.uci.ics.textdb.sandbox;
+package edu.uci.ics.textdb.sandbox.helloworldexamples;
 
 /**
  * Hello world!
