+++ textdb/textdb-sandbox/src/main/java/edu/uci/ics/textdb/sandbox/helloworldexamples/FlavioHelloWorld.java
@@ -1,13 +1,14 @@
-package edu.uci.ics.textdb.sandbox;
+package edu.uci.ics.textdb.sandbox.helloworldexamples;
 
 /**
  * Hello world!
  *
  */
-public class SudeepHelloWorld
+public class FlavioHelloWorld
 {
     public static void main( String[] args )
     {
+        //Just print Hello world
         System.out.println( "Hello World!" );
     }
 }
