+++ textdb/textdb-sandbox/src/main/java/edu/uci/ics/textdb/sandbox/helloworldexamples/ShuyingHelloWorld.java
@@ -1,14 +1,13 @@
-package edu.uci.ics.textdb.sandbox;
+package edu.uci.ics.textdb.sandbox.helloworldexamples;
 
 /**
  * Hello world!
  *
  */
-public class FlavioHelloWorld
+public class ShuyingHelloWorld
 {
     public static void main( String[] args )
     {
-        //Just print Hello world
         System.out.println( "Hello World!" );
     }
 }
