+++ textdb/textdb-sandbox/src/main/java/edu/uci/ics/textdb/sandbox/helloworldexamples/SudeepHelloWorld.java
@@ -1,15 +1,13 @@
-package edu.uci.ics.textdb.sandbox;
+package edu.uci.ics.textdb.sandbox.helloworldexamples;
 
 /**
  * Hello world!
  *
  */
-public class YangHelloWorld
+public class SudeepHelloWorld
 {
     public static void main( String[] args )
     {
-         // Print 'Hello World'
         System.out.println( "Hello World!" );
-        
     }
 }
