+++ textdb/textdb-sandbox/src/main/java/edu/uci/ics/textdb/sandbox/helloworldexamples/YangHelloWorld.java
@@ -1,12 +1,15 @@
-package edu.uci.ics.textdb.sandbox;
+package edu.uci.ics.textdb.sandbox.helloworldexamples;
+
 /**
  * Hello world!
  *
  */
-public class JinggangHelloWorld
+public class YangHelloWorld
 {
     public static void main( String[] args )
     {
+         // Print 'Hello World'
         System.out.println( "Hello World!" );
+        
     }
 }
